refactor(LadderGame): 게임결과 반환 타입 변경
